requirements.txt
crewai
langchain-openai
langchain-community==0.3.24
python-dotenv
